requirements.txt
requests
python-dotenv
pandas
numpy
openpyxl
pywin32